+++ src/panel_material_ui/widgets/__init__.py
@@ -6,5 +6,3 @@ from .menus import *  # noqa
 from .misc import *  # noqa
 from .select import *  # noqa
 from .sliders import *  # noqa
-from .SelectPicker import *  # noqa
-from .SelectSearch import *  # noqa
